Restruct code into separate files
